% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{request_groups}
\alias{request_groups}
\title{Request groups}
\usage{
request_groups(id, regrex, member, signature, limit, offset)
}
\arguments{
\item{id}{A string.}

\item{regrex}{A string}

\item{member}{A string}

\item{signature}{A string}

\item{limit}{An integer}

\item{offset}{An integer}
}
\value{
An S3 object of class openreview.
\describe{
    \item{url}{A string of the url part of the request}
    \item{response}{A list. The return of the GET call}
    \item{content}{A list. The return of the content call}
}
}
\description{
Request groups
}
\examples{
request_groups(id = 'mahshaaban@gnu.ac.kr')

}
