% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{request_invitations}
\alias{request_invitations}
\title{Request invitations}
\usage{
request_invitations(
  id,
  regrex,
  replyForum,
  replyInvitation,
  minduedate,
  tags,
  limit,
  offset
)
}
\arguments{
\item{id}{A string.}

\item{regrex}{A string}

\item{replyForum}{A string}

\item{replyInvitation}{A string}

\item{minduedate}{A string}

\item{tags}{A logical}

\item{limit}{An integer}

\item{offset}{An integer}
}
\value{
An S3 object of class openreview.
\describe{
    \item{url}{A string of the url part of the request}
    \item{response}{A list. The return of the GET call}
    \item{content}{A list. The return of the content call}
}
}
\description{
Request invitations
}
\examples{
request_invitations(id = 'mahshaaban@gnu.ac.kr')

}
