% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{request_notes}
\alias{request_notes}
\title{Request notes}
\usage{
request_notes(
  id,
  invitation,
  forum,
  number,
  trash,
  content.authorids,
  content.pdf,
  content.paperhash,
  details,
  limit,
  offset
)
}
\arguments{
\item{id}{A string.}

\item{invitation}{A string}

\item{forum}{A string}

\item{number}{A string}

\item{trash}{A logical}

\item{content.authorids}{A string}

\item{content.pdf}{A string}

\item{content.paperhash}{A string}

\item{details}{A string}

\item{limit}{An integer}

\item{offset}{An integer}
}
\value{
An S3 object of class openreview.
\describe{
    \item{url}{A string of the url part of the request}
    \item{response}{A list. The return of the GET call}
    \item{content}{A list. The return of the content call}
}
}
\description{
Request notes
}
\examples{
request_notes(id = 'mahshaaban@gnu.ac.kr')

}
