% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{request_profiles}
\alias{request_profiles}
\title{Request a profile}
\usage{
request_profiles(id, email, first, middle, last, group)
}
\arguments{
\item{id}{A string.}

\item{email}{A string}

\item{first}{A string}

\item{middle}{A string}

\item{last}{A string}

\item{group}{A string}
}
\value{
An S3 object of class openreview.
\describe{
    \item{url}{A string of the url part of the request}
    \item{response}{A list. The return of the GET call}
    \item{content}{A list. The return of the content call}
}
}
\description{
Request a profile
}
\examples{
request_profiles(email = 'mahshaaban@gnu.ac.kr')

}
