% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{request_tags}
\alias{request_tags}
\title{Request tags}
\usage{
request_tags(id, invitation, forum, replyto, deleted, limit, offset)
}
\arguments{
\item{id}{A string.}

\item{invitation}{A string}

\item{forum}{A string}

\item{replyto}{A string}

\item{deleted}{A logical}

\item{limit}{An integer}

\item{offset}{An integer}
}
\value{
An S3 object of class openreview.
\describe{
    \item{url}{A string of the url part of the request}
    \item{response}{A list. The return of the GET call}
    \item{content}{A list. The return of the content call}
}
}
\description{
Request tags
}
\examples{
request_tags(id = 'mahshaaban@gnu.ac.kr')

}
