% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{request_users}
\alias{request_users}
\title{Request personal profile}
\usage{
request_users(id, password)
}
\arguments{
\item{id}{A string.}

\item{password}{A string}
}
\value{
An S3 object of class openreview.
\describe{
    \item{url}{A string of the url part of the request}
    \item{response}{A list. The return of the POST call}
    \item{content}{A list. The return of the content call}
}
}
\description{
Request personal profile
}
